+++ src/main/java/com/app/repositories/SectorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.models.Sector;
 
-public interface SectorRepository extends JpaRepository<Sector, Integer> {
+public interface SectorRepository extends JpaRepository<Sector, Long> {
 
 	Sector findByName(String name);
 }
